CIに各デモの静的解析ツールを追加 (#59)
* Python版CIにruffによる静的解析を追加

python-demo / gradio-ver / nicegui-ver それぞれにruffをdev依存として追加し、
CIの各ジョブでpytest実行前にruff checkを走らせるようにした。READMEも更新。


* 未対応言語のCIに静的解析ツールを追加

typescript-demo (deno lint) / csharp-demo (dotnet format --verify-no-changes) /
kotlin-demo (ktlint) / php-demo (PHPStan) にそれぞれ静的解析を導入し、CIの各
ジョブでテスト実行前にチェックするようにした。

- kotlin-demo: ktlintの指摘に合わせてApplicationTest.ktをフォーマット
- php-demo: PHPStan(level6, src/tests対象)の型チェックに合わせてCounterStore::get()
  とRouter::file()の戻り値の型を安全にした（挙動は変更なし）

READMEにも各デモの静的解析コマンドとCIの説明を追記。


---------

+++ php-demo/src/CounterStore.php
@@ -13,7 +13,7 @@ final class CounterStore
 {
     public function get(): int
     {
-        return $_SESSION['count'] ?? 0;
+        return (int) ($_SESSION['count'] ?? 0);
     }
 
     public function increment(): int
